wrolpi/plugins/videos/requirements.txt
youtube-dl>=2019.8.13
webvtt-py==0.4.2